Optimize Queue: avoid most linear traversals
Promised are internally created in class Queue to block a take or an
offer that cannot be immediately satisfied. They are added to a
scala.collection's Queue. A bracket ensure removal from the queue,
however, the promise is never in the queue when it is completed,
so the O(n) removal is necessary only when the promise was interrupted.
We check that this is the case before removing.