Scheduling Adding functions enabling an effect to run only at a given point in time. The functions `loop` and `await` call each other to establish when an effect must run.
The `loop` function calls `foldM` to execute the effect and retrieve its result. At this point, it gives control to `await`

The `await` function keeps passing the last result to schedule and awaits when schedule signal when it's ready to run. Schedule will always receive the last result as parameter to compute the next state.
When schedule is ready to run, `await` return control to `loop`, restarting the process

Added a new flag to Decision to signal when a schedule is ready to run, not only if it must continue or stop