Fix remainder handling in ZStream#transduce (#678)
If the sink indicates done, only run the sink again if there is
a remainder to pass to it.

In the case of the sink requesting more, the extract has to be
done after the upstream fold completes, so it can't be done inside
`feed`. A boolean flag indicates when the final extract is needed.