Fix remainder handling in ZStream#transduce (#678) (#682)
* Fix remainder handling in ZStream#transduce (#678)

If the sink indicates done, only run the sink again if there is
a remainder to pass to it.

In the case of the sink requesting more, the extract has to be
done after the upstream fold completes, so it can't be done inside
`feed`. A boolean flag indicates when the final extract is needed.

* Extend test cases for transduce a little.

* Re-initialize the sink state when there's no remainder.

The earlier fix on this branch broke this case. We should
always call `initial` after calling `extract`. If the upstream has
no more elements, the call to `initial` isn't needed,
but it does no harm.