Use URIO in type signatures where possible
Replace `ZIO[x, Nothing, y]` with `URIO[x, y]`.
IMHO this would improve type inference, albeit in some edge cases.
So I'm not sure if we should do this but I don't see why not.

One edge case is resolving syntax for `ArrowChoice[URIO]` from cats.
Because of partial unification `ZIO[*, E, *]` is not an inferrable type.
On the other hand `ZIO[R, *, *]` remains inferrable with `URIO` as well.