Fix flaky test
`ZIO.forkAll`, when being executed with `List(ZIO.dieNow(boom))` will
internally execute `nilIO.zipWithPar(boomIO)((as, a) => a :: as)`, where
`nilIO` is a simple ZIO value that results in `Nil`, and `boomIO` a
simple ZIO value that results in `ZIO.dieNow(boom)`. In rare cases,
`boomIO` completes before `nilIO` and then interrupts `nilIO`. The
result of the test is then not just a `Die`, but also an interruption.

Fixes #2096