Adapt implementation of `ZIO.forkAll`
`ZIO.forkAll`, when being executed with `List(io)` internally executed
`nilIO.zipWithPar(eff)((as, a) => a :: as)`, where `nilIO` is a simple
ZIO value that results in `Nil`, and `eff` is a ZIO value that wraps
`io`. Although unlikely, `eff` could complete before `nilIO`, and in the
case of an error, interrupt it. This interruption could then be observed
by the caller, leading to somewhat counterintuitive behaviour.

With this change, `ZIO.forkAll` uses special code to handle iterables
containing only one element.

Fixes #2842