Fix flaky test (#2951)
* Fix flaky test

`ZIO.forkAll`, when being executed with `List(ZIO.dieNow(boom))` will
internally execute `nilIO.zipWithPar(boomIO)((as, a) => a :: as)`, where
`nilIO` is a simple ZIO value that results in `Nil`, and `boomIO` a
simple ZIO value that results in `ZIO.dieNow(boom)`. In rare cases,
`boomIO` completes before `nilIO` and then interrupts `nilIO`. The
result of the test is then not just a `Die`, but also an interruption.

Fixes #2096

* Adapt implementation of `ZIO.forkAll`

`ZIO.forkAll`, when being executed with `List(io)` internally executed
`nilIO.zipWithPar(eff)((as, a) => a :: as)`, where `nilIO` is a simple
ZIO value that results in `Nil`, and `eff` is a ZIO value that wraps
`io`. Although unlikely, `eff` could complete before `nilIO`, and in the
case of an error, interrupt it. This interruption could then be observed
by the caller, leading to somewhat counterintuitive behaviour.

With this change, `ZIO.forkAll` uses special code to handle iterables
containing only one element.

Fixes #2842

* Refactor `ZIO.forkAll` to be more compact