use infinity in ZManaged.useForever to avoid GC related problems
described in release notes here (see "New ZIO.never" section):
https://github.com/zio/zio/releases/tag/v1.0.0-RC18

Excerpt:
When ZIO.never is forked, it will be garbage collected, so ZIO.never.onInterrupt(putStrLn("Bye")) will never execute the finalizer. This can be inconvenient for testing, so ZIO.infinity is added which is like never, but won't be garbage collected.