Add isNative to TestAspect, TestPlatform, and TestAspectSpec
This PR adds code for ScalaNative builds which corresponds to the
existing isJVM & isJS code.  ZIO on ScalaNative is experimental and
ZIOFramework is highly experimental.

This change has two benefits. It is small enough to allow easy
review and be non-controversial.  The second benefit is that it
reduces the effort to work on getting ZIOFramework to work on
ScalaNative.

Tested for safety (i.e. did no break existing JVM & JS) by
manually running testJS & testJVM (latter had a few problems with
-Xfatal-warnings, but I think this PR does not cause those.)

I will be testing for efficacy  (i.e. does it do what it says it does)
during my continuing ZIO ScalaNative work.