Add documentation describing ZIO equivalents for Akka operators
At the moment, it's only for Source, and only for operators listed on
akka's main page. If this proves useful, we can extend it for other
streaming data types, and/or further operators.

The reverse could also be envisioned (akka equivalents for ZIO
operators), but there may be less demand for that.