Add includeCase parameter to SummaryBuilder
Suppress the second Cause during ZTestFramework output.

Currently, the output is of the structure:

    - SuiteName
      - Test label
      Fiber failed.
      <snip>
    Ran 1 test in 2 s 167 ms: 0 succeeded, 0 ignored, 1 failed
    - SuiteName
      - Test label
      Fiber failed.
      <snip>
    Done

This obscures the individual test failures, the summary at the very end
should only have the SuiteName and Test label, with stacktrace
suppressed, to highlight which tests are having a problem to make it
easier to scroll up in the test log.

With this alteration, the output of the test structure now looks like:

    - SuiteName
      - Test label
      Fiber failed.
      <snip>
    Ran 1 test in 2 s 167 ms: 0 succeeded, 0 ignored, 1 failed
    - SuiteName
      - Test label
    Done

(notice, the second `Fiber failed.` and stacktrace have been omitted)