Make TestAspect.fibers public
It's required for the proper functioning of `TestClock.adjust`
but we can't apply it in `AbstractRunnableSpec` because it doesn't
constrain the environment in any way (we need `Annotations`).

For projects that extend `RunnableSpec` to customize their
test environment and test runner, currently the only option to make
`TestClock.adjust` work is to hack around the access restriction.