Add sbt-version-policy
Initially, my goal was to declare the versioning scheme used by zio. sbt-version-policy does that, and it also checks that your version numbers match the declared versioning scheme.

The benefit of declaring the versioning scheme used by zio is that all the users of zio will benefit from accurate diagnostics of conflicting dependencies.