Fix race condition in fromOutputStreamWriter (#5872)
Because the input and output streams are piped together,
when the output stream is closed, the input stream completes.
Due to the semantics of `drainFork` that would sometimes cause
the output stream writer to be interrupted before completing
the error channel promise. But in fact we don't need a separate
error channel, we just need to make sure the input stream reader
waits for the output stream writer to complete.