Report error messages to sbt events which end up in JUnit XML
This is a quick and dirty solution reusing `DefaultTestReporter`
and subsequently stripping the ANSI colors from the result.

A more elegant solution would involve writing a new reporter,
or adding an option to render without colors to `DefaultTestReporter`,
but this is probably a major issue for anyone running ZIO test in CI,
so I think the quick win is worth it.