MockableMacro: fix warning on produced code
Most of the time, code generated by MockableMacro make scalac
complain about `rts` not being used.
This is a big problem when `-Xfatal-warnings` is used. One could
say that the solution will be to use `-Ywarn-macros:before` but some
project need or at least recommand usig `-Wmacros:after`, for example:

https://diffx-scala.readthedocs.io/en/latest/#tips-and-tricks