Fixes grammatical errors and edits code snippets (#6667)
Updated some of the code snippets to reflect organizational best practices and removed the last example method in the retry section: retryOrElseEither. I made this choice because, unlike the other methods, there was not a code example and creating an easily understandable one which built upon the previous examples proved difficult and would have likely been more confusing than informative. Additionally, it seems that maybe this method is not a common go-to.