PR change request.  'launch' semantics have changed.
The signature of 'launch' has been changed so the provided Schedule is now 'Any' instead of 'Out'.

With this change along with a change to the way the Schedule is applied to the underlying effect, the semantics of 'launch' have  changed as well.

Before, the underlying polling effect would always run once, regardless of the Schedule.

Now, the polling behavior is more intuitive in that it matches the provided Schedule.

Before:
'Schedule.recurs(1)'  would result in the polling effect being called twice.  Once to get the base metric ouput and then once more to satisfy 'recurs(1)'.

After:
'Schedule.recurs(1)'  would result in the polling effect being called only once.

The only drawback here is that the Schedule can no longer use the Metric's value.