Optimise `_IdentityFn`
Here's how I see this (and I can be wrong):
The probability that someone using ZIO doesn't use this `_IdentityFn` lazy val is IMO close to 0.
Each time you flatten a ZIO, you probably use it.
AFAIR, using a `lazy val` over a `val` adds several instructions in the bytecode to make the `lazy` mechanism work.

Because, this probably is low, the memory impact of this `val` not being lazy is extremelly low and the additional bytecode instructions needed for the `lazy` mechanism, I think we can just remove the `lazy` part

WDYT?