Extend causes in the Exit.getOrThrow* methods with stack traces
The idea is that most users call those methods in interop code
and since they are leaving the ZIO runtime, they need a stack trace.

Also override `printStackTrace` in `FiberFailure` by delegating
to a new `Cause.prettyPrintWith` method which allows for a custom
`println` function. This is necessary because the original method
uses the private `stackTrace` field, not `getStackTrace`.

Updated `Cause` to remove the "Exception in thread" part as this
shouldn't be part of the exception `toString`, it's added later and
was causing it to be printed twice.

Last thing to node is that I didn't change `FiberFailure.toString`
for caution of breaking user code but it's not in line with Java
exceptions, which don't include the stack trace in their `toString`.