Optimize `TestArrow.run` to double stack safety
This change roughly doubles the stack safety of TestArrow.run by removing indirection with the attempt method on each iteration of `run` loop.

I split out the `run` method to an inner `runLoop` method and a top level `run` that uses try/catch once. This will prevent method indirection and the allocation of a by-name parameter.